requirements.txt
numpy~=1.18.5
matplotlib~=3.2.2
tensorflow~=2.1.0
setuptools~=47.3.1
pillow~=7.1.2